Add support for multi-device runs
Currently, there can be at most one outstanding session (on one or more devices)
at a time (per Run/Debug). However, with the new device selection drop down UI,
it makes sense to support at least multiple concurrent Run sessions on different
devices. This CL adds support to have multiple tabs in the Run tool window, with
each tab corresponding to a different device.

Note that device names displayed directly in the tabs aren't supported yet (as
this will be much more involved), and instead the device name (and in the future
the snapshot name) is printed to the console of the corresponding Run tab.

Change-Id: I11c1eb9e3f80eb2038f1e85b7162afdd27526cd7
Fixes: 122726017
Test: N/A