Fix resource item name not changing
Look for old and new name when removing the ResourceItemSource of a
renamed file because the file name might already have changed when
reaching the PsiListener callback so no ResourceItemSource with the
oldName will be found and thus be removed from the repository resulting
in a PsiResourceItem pointing to the good file but with the old name.

This only happens when the PsiResourceItem has been created so we need
to perform the rename twice to test this behavior.

Test: Add test for renaming a file twice
Bug: 123213449
Change-Id: I9911028e37c306ac7144b0bd0f708a55c0257c11