Refactor ProfilerService's API usage on Studio-side to TransportService
- Moved existing ProfilerServiceProxy logic to TransportServiceProxy
- Moved all device/process/agent logic from ProfilerTable to UnifiedEventsTable
- Datastore now intiiates the connection with perfd via the new TransportService
- All test infrastructure now initializes the fake GrpcServer with both the
FakeTransportService and FakeProfilerService. (As long as StudioProfilers is
involved, we need the TransportService for Device/Process logic and ProfilerService
for sessions-related logic).
- FakeTransportService and FakeProfilerService now shares the same FakeTimer to
make tests passed in both the new and old data pipeline. A few if-else had to be
added in some tests to check whether an event has occured in the FakeTransportService
or the FakeProfilerService. In particular, sessions are managed via the generic
TransportService in places where we test the new pipeline, as opposed to the
BeginSessionRequest/EndSessionRequest in the old flow.

Bug: 73538507
Test: Existing and migrated

Change-Id: I9b51abb155b5d65b7335760404af849ccda086fb