First checkpoint for transport pipeline demo
New internal action for setting up a minimalist demo of the transport pipeline
used by the profilers. The action shows a dialog that allows users to connect
to an app, send commands and observe the events that are sent back to Studio.

Majority of the code is copied and pasted over from StudioProfilers
and SessionsView.

Bug: N/A
Test: N/A. Internal demo only

Change-Id: Ia7ef3e72a384108d83a97d17784eb87600a10d74