Avoid calculating actions to display dropdown arrow
DropDownActionButton would need the full actions list in order to decide
when to display the drop down arrow. In many cases however, this should
not be necessary since we know that we need to always display it. This
allows delaying the cost of the loading of the actions in the cases
where it's more expensive so we only do it when the user interacts with
it.

Bug: 119688277
Test: Existing tests pass
Change-Id: I5aeb043c4a7f7e333c455311d85fa16e5b1549f3