Fix PsiInvalidElementAccessException on NlComponent#setAttribute
Follow up from the revert 6281430.

The previous fix also had an issue where one of the code calling
setAttribute was done outside WriteActionCommand, which should have
failed but it probably relied on the behaviour that new value was
the same as the existing one.

This fix ensures that processes are correct, but does not protect from
IncorrectOperationException, which will be thrown when setting attribute
outside the WriteCommandAction.

Bug: 123704869
Test: Unit test + drag n drop works

Change-Id: Ib239a0fe53a3d83a3f0ae922e6ee086008d3ce0a