Clean up agent status timeout logic
This change is a follow-up to a1ac328d6d1ba7487d0ca34289b7e2924537d98d.
We now cache the retry count so that switching in and out of a session
that is loading does not reset the retry count. This also fixes an odd
behavior where clicking on the session will cause the UI to jump from the
loading phase to the overview stage immediately.

Bug: 124254814
Test: Added unit tests

Change-Id: Icf4e90695965e96c8016ac76bd13522cd282bef9