Fix bleeding control background colors in property panel.
Reason for this change:
All of the controls are used in a toolwindow, which is using
secondaryBackground. Without it you see funny bleedthrough and
strange background effects. The color constant used will change
appropriately for high contrast LAF.

You could argue that the background should be configurable.
But all our uses of the properties panel appear in toolwindows
with that background. So the background is hardcoded for now.

Bug: 124527812
Test: N/A
Change-Id: I3c8905085f614d22fa8cac17697d1cbbd632f465