Fixes issue where EDT was being stalled for 500ms when devices are plugged in
This CL fixes an issue where the Apply Changes/Apply Code Changes's base action
was attempting to resolve the AndroidVersion of the underlying device on the
EDT when the device is plugged in, causing a 500ms stall.

The fix is to move the version query into the DeploymentApplicationService,
which is the new home/layer to fix some of the existing issues of ddmlib. The
getVersion call in the new layer is now a Future that will resolve to the
AndroidVersion when done.

Fixes: 124465261
Test: manually tested
Change-Id: Ifc89215b731215adfd8cb06e2c0d9d8fec8dd153