Show a warning event log when JDK is not Java Home
We now show an option to setup JAVA_HOME as JDK location when Android
Studio is run for first time but if the user decides to copy settings
from a previous installation then this is not seen by them.

With this CL, everytime that sync finishes (successfully or with errors)
we check if JDK is set to JAVA_HOME and if that is not the case then we
show a warning in the Event log with two quickfixes: one to set
JAVA_HOME as JDK and other to not show this warning again. Also made
some small cleanups on related classes and changed order in which
suggest fixes to preffer JAVA_HOME over embedded JDK.

Sreenshot with message:
https://screenshot.googleplex.com/H4MsyBEmRQ3

Bug: 115925397
Test: Added test for new hyperlinks and updated existing tests.
Change-Id: Ia812fb7945b4918bb5a7a21bafccbe895526531f