Do not pass the map to the virtual delegate
KeyToConnectionTimeMap is a map of device keys to the times their physical
devices were connected or their virtual devices were launched. We treat the
virtual devices built by the virtual devices worker delegate as disconnected
so it never made sense to pass the map to the builder. Passing the map at
that point gave a worker thread access to it which interfered with the
sensible access by the event dispatch thread, causing concurrent
modification exceptions.

Bug: 125517216
Test: Not applicable
Change-Id: Ia83d7fa4e61bf291dc239b6f2302d8068bf47a3a