Ignore disposed modules in various icon providers and decorators.
DirectoryIndex lags behind the VFS state and may return a disposed
module. This usually is not a problem unless a module is disposed just
after is has been created or updated. This should not generally happen
there are cases where it might be unavoidable or difficult to eliminate
in the foreseen future.

This CL makes UI to ignore disposed modules when fetching various icons.

Bug: 126802443
Test: existing (fewer UI tests should be flaky)
Change-Id: I01264532cd171f8988bb7d35bf7a27bfb205bf79