Converting views should not be done in the UI thread
MorphComponentAction is used to convert views (e.g. TextView into a Button)
and might eventually call NlDependencyManager#addDependencies to cover the
case of converting views into components with missing dependencies. As this
method is blocking, we shouldn't be calling it from EDT.

This CL changes the view conversion to be done in a separate thread backed
by a ProgressIndicator.

Bug: 127249013
Test: NlEditorTest#morphVieGroup and NlEditorTest#morphComponent

Change-Id: I2a8cd3ea96387e9ee2c2da53afa2444486585cfe