Avoid holding on to dynamic proxies in JavaModuleModel
IdeaCompilerOutputImpl is basically a data class holding the names of
output folders. Instead of holding on to dynamic proxies, which use
org.gradle.tooling.internal.adapter.ProtocolToModelAdapter$InvocationHandlerImpl
in turn pulling ProtocolToModelAdapter$ViewGraphDetails, which has a cache
of other dynamic proxies, etc., we just copy the four files.

This seems to be the last codepath from the old sync which didn't
perform a deep copy.

See also JavaModuleModelFactory.getCompilerOutput().

Bug: 117274283
Test: existing
Change-Id: I6defcd7d6e5ab6e263a7faa3c3f53fdd684cdb89