Work around Linux clipboard issue in AddKotlinTest
The Linux clipboard issue is known to be unreliable, as discovered in
change I10cd4530bb3ae25ee6ce5266eec626ec1f99478d. As a quick summary,
values stored into the clipboard are not immediately available for
reading afterwards. Change I10cd4530bb3ae25ee6ce5266eec626ec1f99478d
addressed this problem by having a special Robot wait for the value to
be in the clipboard before reading from the clipboard.

The I10cd4530bb3ae25ee6ce5266eec626ec1f99478d workaround works by
polling the clipboard to check if the clipboard passes each of the
following tests:

1. Does the clipboard contain the string "data flavor"?
2. Are the clipboard's contents are equal to the value that was last
stored into the clipboard?

However, the workaround added in
I10cd4530bb3ae25ee6ce5266eec626ec1f99478d is not foolproof. If the
clipboard is used multiple times, as it is here in AddKotlinTest, then
the IDE can still fail to read from the clipboard in some cases. I
currently do not know why, but I guess that when a new value is stored
into the clipboard, the clipboard begins updating the list of available
data flavors and stores the new value into the clipboard non-atomically.
That means that it is possible to observe the clipboard where the list
of available data flavors do not describe the value stored in the
clipboard.

For example, suppose the clipboard has made the newly stored value
available for reading, but the clipboard is still in the process of
updating the list of available data flavors. In this case, the 2 tests
of the workaround would pass, and the test would proceed with running a
paste action.

The IDE handles the paste action by going through its own set of
internal clipboard checks. The IDE's clipboard handler checks the
clipboard to see if a list of available data flavors contains the string
data flavor. At this point, it's possible the clipboard has started
working on updating the list of available data flavors, but the list of
available data flavors does not yet contain the string data flavor. The
IDE's clipboard handler sees that no string data flavor is available,
and the paste operation leads to nothing happening in the IDE. This
would leave the text field empty, and the test would fail due to the
text box not containing the value that's expected to be in the text box.

There does not seem to be a way to ensure that the clipboard has
completely finished updating its internal state after storing a value
into the clipboard in a test. Due to this reason, this change addresses
this problem by having the NewKotlinClassDialogFixture type in the name
of the new Kotlin file rather than by pasting the name. This does slow
down the test, but this is acceptable as sanity tests are not too
concerned about speed of the test.

Bug: 125934368
Test: bazel test //tools/adt/idea/android-uitests:AddKotlinTest
Change-Id: I1e63dffa71a51ddaa95d373817ecdf78b8eac374