Adding a zoom panel over the design surface
Adds a floating zoom panel on the lower right corner of the design
surface.

Allows to zoom in, zoom out, zoom to 100% or fit to scren, similar to
the toolbar zoom buttons.

The label for the current zoom value is a panel that is visible for a
short period of time when the zoom changes. May stay visible as long as
the panel is interacted with.

Guarded under new flag: NELE_DESIGN_SURFACE_ZOOM. Turning it on will
substitute all zoom controls from the action toolbar to the
DesignSurface. With some exceptions like the Intellij ImageViewer,
Layout Inspector and NavEditor right click zoom controls.

Screenshots in bug.

Bug:126902858
Test: manually, unit tests for zoom and pan actions

Change-Id: I8c03340cd5ad12039b19352a9085fecf8b5dddf8