Add memory leak test for inspections
Runs inspections (Analyze > Inspect Code...) on the whole project, then
closes the inspections tool window.

The implementation of ToolWindowFixture.close() is somewhat unfortunate,
but clicking on the X on the tabs via the UI is difficult since the X
is an icon inside a label, not a separate component. Furthermore, to
access this icon (to determine its location) would require either
reflectively accessing or widening the accessibility of both the
ContentTabLabel class and the myAdditionalIcons field therein.

Bug: 124256751
Bug: 127830884
Change-Id: I4b03f65d46119d59c507f26dadd9ebd9b9981902