Fix thread problem while displaying completions.
The new properties panel allows the completions to be computed on
a pooled thread. There currently is a bug where it is possible that
the background thread will update the ui component for the
completions while they are being used by the ui-thread.

Fix this by updating theui components on the ui thread.

Fixes: 128284728
Test: N/A
Change-Id: I6a541ef58c38a281bcea158142e0f037d60ee41c