[Gradle Sync] Ensure test failure it correctly detected
It is possible for syncSucceeded to be called before syncFailed if the
test fails with SyncIssues. We fail this be throwing an illegal state
exception in PostSyncProjectSetup#failTestsIfSyncIssuesPresent, this is
called after syncSucceeded and triggers syncFailed from the thrown
exception.

This change adjusts the sync listener to account for this by setting
success to false when syncFailed is called. This is a temporary measure
so that tests can reliably detect test failures.

Bug: N/A
Test: N/A
Change-Id: I7ea3c25903d950b45f0834d4114579cca6e05369