[Gradle Sync] Don't overwrite sdk/jdk for tests
It's currently impossible for tests to overwrite the default sdk/jdk
location to an invalid one due to these PreSyncChecks that always find
studios SDK and sync the LocalProperties with it.

This change allows the sync Request to skip the pre-sync checks if
needed. We also add some wiring in AndroidGradleTestCase in order to
allow tests to customise their Request when syncing.

Bug: N/A
Test: N/A
Change-Id: I4bed1c17365104339a9f9c312305d971025a4a30