Use the resolved dependencies when adding a dependency.
When we are adding a dependency by dropping a component from the
palette, we were attempting to add the new dependency if it wasn't
already an explicit dependency in the module.

Unfortunately the information about the direct dependencies were
flawed - we only have access to the dependencies we can parse
ourselves from the gradle file. And sometimes we cannot parse the
dependency section at all.

With this CL we temporarily change the default to only add a new
dependency if it was missing in the transitive closure of the
module dependencies.

Fixes: 123222832
Fixes: 112716902
Test: Changed the tests to respect the output from the resolved
      dependencies.
Change-Id: I662c7cc8e488659d98b39b37a1ddc3fb756c425c