Back MergedManifestManager with AsyncSupplier
MergedManifestManager provides APIs for obtaining a cached snapshot
of the merged manifest if available and recomputing the merged manifest
when the cached snapshot has become stale. Currently, the getSnapshot()
API will block the calling thread if the merged manifest needs to be
recomputed, and this design makes it really easy for callers to
accidentally cause UI freezes by calling the method on the EDT.

This CL resolves the issue by making MergedManifestManager a module
service that exposes a MergedManifestSupplier (a new implementation
of AsyncSupplier<MergedManifestSnapshot>). Because we expect the merged
manifest snapshot to be requested and invalidated frequently, the
supplier delegates to a ThrottlingAsyncSupplier. The delegate supplier
not only offloads the heavy lifting to a background thread and provides
a future that callers can respond to asynchronously when the new manifest
snapshot is ready, but it also throttles the recomputation rate so that
we don't waste resources by constantly recomputing stale snapshots as
the user edits a manifest.

The existing static utilities getSnapshot() and getCachedSnapshot()
have been deprecated in favor of the AsyncSupplier API, which they
now also use. Unfortunately, there are legacy callers who expect
to block on creating the merged manifest while also holding the
global read/write lock. Blocking on a background thread in such cases
would lead to deadlock since creating the merged manifest requires the
read lock, and so we're forced to handle the special case by doing the
heavy lifting in the calling thread outside of the
ThrottlingAsyncSupplier delegate.

While MergedManifestSupplier still handles caching the snapshot created
in such cases, this adds some complexity to the code. I've tried to
mitigate some of the added bloat by migrating the code to Kotlin.

Bug: 128610392
Test: MergedManifestManagerTest
Change-Id: Ibab7329039244b7139792151acbfe5f17cd95849
Test-Strategy: ui.test.groups=DEFAULT,EDITING