Use dropdowns for PSD JDK and NDK locations
This is a change in UI, now we use dropdown boxes to suggest known
setting for JDK and NDK paths instead of having radio buttons for JDK
and a link for NDK. The new UI is more clear on what it does. Also the
order in which the options are presented changed to have SDK and NDK
together.

The combo boxes are also editable and the values are validated when
pressing Ok or Apply. Validation is also done when selecting the paths
using the browse buttons.

Everytime the value of the JDK como box is changed, we check if it is
not the same as JAVA_HOME and in that case display a message saying
this:

"To use the same Gradle daemon between Android Studio and the command
line, select JAVA_HOME from the drop-down"

Bug: 115925397
Bug: 123246879
Test: Current UI tests passes.
Test-Strategy: ui.test.groups=PROJECT_SUPPORT,PROJECT_WIZARD,EDITING,TEST_FRAMEWORK,DEFAULT
Change-Id: I243b28661d0eb853836bd322faca24e0fb3f42e9