Fix StackOverflowException when starting logcat.
Set preferred focusable component in tool window to avoid bug in
ContentImpl.

The problem in ContentImpl.getPreferredFocusableComponent is that
if no preferred focusable component is set then it uses the
focus cycle root ancestor to determine the preferred component
to gain focus when the tool window is opened. That ancesestor is
typically going to ask the components in its content causing an
infinite loop.

Fixes: 129117429
Test: N/A
Change-Id: Ic81429ec256d5ad15e5e7a3e71aca181f94c55ed