Add scrollToCenter behavior for the layout editor
When selecting components (wether it is on the surface or component
tree), we should now scroll the view to focus the selected component but
only when the component was not fully visible in the surface.

Similar to navigation editor, but with no animation/smooth scrolling at
the moment and somewhat different rules.

Bug: 126902858
Test: n/a
Change-Id: I80757b3d320657aeda301dc7e64c228d7cfb27c7