Switch to IDEA sync if buildSrc module exists
buildSrc modules are the modules that exist under root project
folder with name "buildSrc". They are not available in BuildAction
if BuildAction is created with root project.

To support buildSrc modules, Sync needs to make a second connect to
Gradle Tooling API with the path of buildSrc directory.

In IDEA sync, this was done by GradleBuildSrcProjectsResolver, while
new sync doesn't have support of it.

Since adding this support to new sync is not trival, switches to
IDEA sync mode if "buildSrc" module exists before we have an actual
fix in new sync infrastructure.

Bug: 118392606
Bug: 123032843
Test: SingleVariantSyncIntegrationTest, EnableDisableSingleVariantSyncStepTest
Change-Id: Iaf6d54bc8d2d0fdbf517706253e4236af18c08ff