Add protection against exceptions preventing registration with the Disposer
The heap analysis report linked from b/130253739 indicates that a DynamicResourceValueRepository was present in the Disposer tree without a parent. This is only possible if a normal control flow was interrupted between DynamicResourceValueRepository.registerListeners calling the MessageBus.connect method passing "this" as the argument and ModuleResourceRepository.forMainResources registering the repository as a child of a ModuleResourceRepository.

Exceptions in the ModuleResourceRepository.addRepositoriesInReverseOverlayOrder method are possible according to b/130315019. The crash database contains an exception for the same session as the memory leak report but without a stack.

The change adds protection against exceptions in the ModuleResourceRepository.forMainResources and DynamicResourceValueRepository.create methods.

Test: existing
Bug: 130253739
Change-Id: Iafe0a6909fa67a91e0d035928799bbcd6094e54a