Compute the root of the tag rather than the document.
The logic for moving namespaces to the root was always
moving namespaces to the root of the layout rather then the
root of current component. This causes a problem when updating
attributes on an in memory component. We ended up changing the
layout when we "just" wanted to change the in memory components.

The failures seen are writes outside of having the write lock,
which started after the commit of change: ag/7019999

Fixes: 130442920
Test: This fixes UI test
      DragPreferencesFromPaletteToComponentTreeTest.dragAndDrop
Test-Strategy: ui.test.groups=DEFAULT
Change-Id: Ia480b2488419bfe0137873e5016249abd759db02