Show a warning for projects with kts build files
Android Studio is able to open projects with kts Gradle build files but
some functionality is not available for them. This CL causes a warning
to appear every time a project with kts build files is opened or the
first time it is synced and it contains a kts file.

Test: New tests added to confirm warning is shown when appropriate.
Bug: 130226195
Change-Id: Id2e3bd6103364499e8351ca083a623b36fb384f5
(cherry picked from commit 8cee7b24482db3154171be4a2d1b254b8dadb58a)