Show antivirus notification after Gradle builds
The antivirus notification will now be shown only after a successful
Gradle build with mode ASSEMBLE, ASSEMBLE_TRANSLATE, REBUILD, BUNDLE, or
APK_FROM_BUNDLE. Notifications are limited to one per day, though the
checks and metrics are not. The code for showing the notification has
been moved to WindowsPerformanceHintsChecker, as the increasing number
of modifications to AndroidStudioSystemHealthMonitor.showNotification
were becoming unwieldy.

Make the "Don't show again" action on the notification
application-scoped rather than project-scoped, and add a "Don't show
again for this project" action.

Bug: 130436464
Bug: 127513071
Change-Id: Ie083205188f7d9f384605973adb484d8028f48b2