Send finish event after other events are processed
When we detect an older plugin version is used we suggest the user to
upgrade it. We used to do that in dialog that was shown before
finishing sync, but now we show a balloon dialog (unless this balloon
is disable using BALLOON_UPGRADE_NOTIFICATION). After upgrading we
immidiatly send an event to the build view saying that the sync event
finished, but if there are still pending messages, those messages will
cause an assertion.

This CL causes the finish event to be sent using invocaLater, causing
any pending messages to be processed before it.

Test: Manualy confirmed the assertion is not caused and covered in
manual tests as well.
Bug: 120118285
Bug: 128024913
Change-Id: I92655604758134719ebe4a5f675c207c0c19939f