Load merged manifest tab in the background
Currently, the merged manifest tab in the manifest editor will show you
a potentially stale snapshot of the merged manifest, or, if a cached
snapshot isn't available then the editor will freeze the UI while the
merged manifest is being computed.

This CL resolves both issues by updating ManifestEditor to use the new
merged manifest APIs that are backed with an AsyncSupplier. This allows
us to show the user a loading spinner while we compute an up-to-date
merged manifest snapshot and then asynchronously update the UI once the
snapshot is ready.

As a follow-up, we could present the editor using a stale snapshot (if
one is available) with some additional UI component to indicate that
we're computing a fresh version in the background. That way, the user
isn't blocked in case a stale snapshot is available.

Test: ManifestEditorTest
Bug: 128610392
Change-Id: I2074b016c9fbc5b168c10627f16deaf42ac51282