NDK side-by-side: offer hotlink to update ndkVersion in build.gradle
Before, if there was a failure to locate NDK then Android Studio
would offer to update ndk.dir. Now, if there is already an NDK
installed then Android Studio will offer to set android.ndkVersion
to the highest version currently installed. If there is no NDK
installed then we will offer a hyperlink to install the latest
NDK and also set android.ndkVersion to that version.

Not in this CL but considering next:
- If there is no android.ndkVersion set then have a balloon that
  suggests adding one (with hyperlink to do it automatically)
- If there is a more recent NDK available then have a balloon
  with a hyperlink to download and set android.ndkVersion.
  Generally, the goal is to get people away from using ndk.dir and
  to get more people to have android.ndkVersion in their
  build.gradle.

Bug: 130363042
Test: Unitested
Change-Id: I5a76a00d9ff6c944eb19ef45d743d3d5d51461e9