Keep Stop button enabled even if Application did not connect
If the user is using a non-debuggalble app on a non-userdebug phone, then
the Client will never connect. In this case, the Stop Action will be disabled.
We want to keep it enabled so users can stop the process even if the process
never connects, and only disable it if the process comes online then goes
offline without pressing the Stop button.

Fixes: 131157731
Test: N/A
Change-Id: Ie03402d1337ef5543d07a0da3bbc38a865d7d9ec