Show stale merged manifest in editor while recomputing
Currently, if an up-to-date merged manifest snapshot is unavailable
when the user opens the manifest editor, then we unconditionally show
a loading spinner while we recompute the merged manifest in the
backgound.

This CL updates ManifestEditor so that if a stale snapshot has already
been cached then we will show the user that snapshot with an editor
notification to let them know we're recomputing the merged manifest in
the backgound. During this time, the user won't be able to make changes
to the stale merged manifest through the editor. Now we'll only show the
loading spinner if this is the first time the merged manifest is being
computed this session.

Bug: 128610392
Test: ManifestEditorTest still passes
      + manual test with thread-only breakpoints confirms stale
        snapshots get shown
Change-Id: I7f59d26bf247f887c30c6ffa2a41e0d4e96e8a14