Split IdeAndroidProject from gradle tooling model.
The tooling model interfaces have different compatibility requirements,
so splitting them should allow both to evolve separately.

Some of the Android Gradle Plugin APIs implement the tooling model
interfaces, which we don't want, so making this split should enable
starting to untangle that too.

This change only does that for AndroidProject to unblock the next change
to split out build types and product flavors. More changes to follow to
clean the rest up.

Most of the changes are in lint, which carries a copy of the model
around, so now lint in Gradle wraps the model in the IDE class first.

Bug: 131735732
Test: Existing unit and integration tests.
Change-Id: I2e50fb07eb201a5078225a748524fda7636a3813