Always checking draggable in ConstraintLayout
ConstraintLayout checks the constraints to
place the component after dragging. So the
actual position after dragging may be different
than the mouse position.

Before this CL, it checks the position when
Live Rendering is enable only. This CL makes
ConstraintLayout checks the draggable area
no matter Live Rendering is enable or not.

Fixes: 129681462
Test: Updated CommonDragTarget
Change-Id: I3b66edea1b3fd7b9cedeea60592ee6a99bc523cd