Don't block on merged manifest in LaunchCompatibilityChecker
LaunchCompatibilityCheckerImpl uses an Android module's minSdkVersion
to determine whether it's compatible with a given device. It obtains the
minSdkVersion from AndroidModuleInfo#getRuntimeMinSdkVersion, which
blocks on merged manifest computation if the AndroidModel doesn't
specify one and the manifest hasn't been computed for the module yet
this session. Since UI components related to device selection use
LaunchCompatibilityCheckerImpl to signal to the user which devices are
compatible, the end result is UI freezes of 2.8s on average (the actual
time depends on the module and how many dependencies it has).

This CL resolves the issue by migrating
AndroidModuleInfo#getRuntimeMinSdkVersion to use the new asynchronous
API for obtaining MergedManifestSnapshots.
LaunchCompatibilityCheckerImpl now checks to see if the minSdkVersion is
readily available and, if not, just defaults to AndroidVersion.DEFAULT
instead of blocking. We still handle incompatible versions gracefully in
the worst case scenario: if the user tries to deploy the app to a device
whose SDK version is less than the minSdkVersion before we can show the
compatibility warning, they'll just get an error in the run console
telling them the versions are incompatible.

Bug: 132289390
Test: LaunchCompatibilityCheckerTest
Change-Id: Ib6d6d75e60c7be086ae74823d5e477c144b06c1d