Fix selection marquee/box not drawing sometimes
Fixes the case when the selection box is not drawn when the root
component is not selected, since it would trigger a selection and cause
it to not render. Overall, triggering a layout pass (e.g from making
selection changes) may cause targets to not render in Layout Editor.

Note that this change causes the previous selection to not get cleared
while dragging to select in layout editor. Should remain unchanged in Nav editor.

Fixes: 131624502
Test: manual tested for LayoutEditor, NavDesignSurfaceTest passes
Change-Id: Iae1ceb055c57c3527b840ad941ea52ce13e9c2d3