Show UseJavaHomeAsJdkHyperlink only when valid
When we detect that the Jdk is not configured correctly we show a
message and a couple of hyperlinks to help users solve this issue,
giving priority to using JAVA_HOME. We are currently offering it without
checking if the JDK version pointed by JAVA_HOME is valid or not.

This CL also checks that the JDK in JAVA_HOME is a valid JDK before
trying to offer it.

Bug: 124680713
Test: new tests in JdksTest
Change-Id: I30e4a1dd2d1a3993abf370170643d9fdbcd4a2d5