Fix for invalid IJ currently selected configuration state
When the facet is disposed, the "selected configuration" still points to
a now-invalid config with a disposed facet. This causes subsequent calls
to the application ID provider to fail because internally the provider
will attempt to register (in the disposer tree) instantiated variables as
dependents to the already-disposed facet.

The fix simply checks if the facet has already been disposed, and early
exits if so.

Test: N/A
Fixes: 132704978
Change-Id: I219fc33d1b63522079ccaf623490da071e7cab31
(cherry picked from commit d05c626d0a184128b0e7c86e8e3ad3033c9b69bd)